+++ lang/lb.js
@@ -1,0 +1,11 @@
+
+$.fullCalendar.lang("lb", {
+	buttonText: {
+		month: "Mount",
+		week: "Woch",
+		day: "Dag",
+		list: "Terminiwwersiicht"
+	},
+	allDayText: "Ganzen Dag",
+	eventLimitText: "méi"
+});